refactor: simplify apps structure and improve dependency examples
- Remove echo application (hello-world is sufficient for simple app example)
- Rename foo to web-app for clearer purpose as dependency example
- Remove internal/ directory to minimize complexity
- Simplify pkg/ to only contain common package
- Update depends_on from foo to web-app/pkg for more granular dependencies
- Simplify api-server and worker implementations
- Update CLAUDE.md to reflect new structure
