--- apps/echo/go.mod
@@ -1,3 +1,0 @@
-module github.com/yuya-takeyama/monotonix-playground/apps/echo
-
-go 1.24.4
