--- apps/foo/cmd/worker/monotonix.yaml
@@ -1,54 +1,0 @@
-app:
-  name: foo/cmd/worker
-  depends_on:
-    - foo
-jobs:
-  build_prd:
-    on:
-      push:
-        branches:
-          - main
-    configs:
-      docker_build:
-        registry:
-          type: aws
-          aws:
-            iam: prd
-            repository: prd
-        tagging: semver_datetime
-        platforms:
-          - linux/amd64
-      docker_build_go_build:
-        go_version_file: apps/foo/go.mod
-  build_dev_main:
-    on:
-      push:
-        branches:
-          - main
-    configs:
-      docker_build:
-        registry:
-          type: aws
-          aws:
-            iam: dev_main
-            repository: dev_main
-        tagging: always_latest
-        platforms:
-          - linux/amd64
-      docker_build_go_build:
-        go_version_file: apps/foo/go.mod
-  build_dev_pr:
-    on:
-      pull_request_target:
-    configs:
-      docker_build:
-        registry:
-          type: aws
-          aws:
-            iam: dev_pr
-            repository: dev_pr
-        tagging: pull_request
-        platforms:
-          - linux/amd64
-      docker_build_go_build:
-        go_version_file: apps/foo/go.mod
