--- apps/foo/go.mod
@@ -1,3 +1,0 @@
-module github.com/yuya-takeyama/monotonix-playground/apps/foo
-
-go 1.24.4
