+++ apps/web-app/go.mod
@@ -1,0 +1,3 @@
+module github.com/yuya-takeyama/monotonix-playground/apps/web-app
+
+go 1.24.4
